\name{block}
\alias{block}
\title{Block of same initial [Fe/H] for the grid}
\description{
The function computes the group identifier for grid rows, according to their
initial metallicities.
 }
\usage{
    block(grid)
     }
\arguments{
\item{grid}{the grid of stellar models.}
}
\value{

  \code{block} returns a vector with the group identifier for
  each row of the supplied grid. 

}

\keyword{manip}
