\name{SCEPtERbinary-package}
\docType{package}
\alias{SCEPtERbinary-package}
\title{A grid-based estimation tool for the age of stars in bynary systems }
\description{
  The package estimates the age of stars in double-lined detached binary
  systems, given  observational 
  effective temperature, metallicity [Fe/H], mass, and radius of both stars.
  The results are obtained adopting a maximum likelihood
  technique on a grid of computed stellar models.
  
}
\references{
  G. Valle, M. Dell'Omodarme, P.G. Prada Moroni, and S. Degl'Innocenti
(2014). \emph{Uncertainties in grid-based estimates of stellar mass and
  radius. SCEPtER: Stellar CharactEristics Pisa
  Estimation gRid}. A&A, 561, A125.

G. Valle, M. Dell'Omodarme, P.G. Prada Moroni, and S. Degl'Innocenti
(2014). \emph{Uncertainties in asteroseismic grid-based estimates of stellar ages. SCEPtER: Stellar CharactEristics Pisa
  Estimation gRid}. A&A, accepted.

G. Valle, M. Dell'Omodarme, P.G. Prada Moroni, and S. Degl'Innocenti
(2014). \emph{Grid-based estimates of stellar ages in binary
  systems. SCEPtER: Stellar CharactEristics Pisa Estimation gRid}. A&A, submitted.
}
\keyword{package}
